refactor: structure inputs in individual folders, rename Field and FieldWrap helpers
