+++ src/core/Input/Switch/styles.css
