+++ src/core/Input/ToggleField/index.js
@@ -2,7 +2,9 @@
 
 import React from 'react'
 import PropTypes from 'prop-types'
+
 import { bemClassNames } from '../../../utils'
+
 import './styles.css'
 
 const bem = bemClassNames('toggle-field')
